+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/NewEnchantmentsFeature.java
@@ -12,9 +12,23 @@ import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.insanelib.event.HurtItemStackEvent;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.stats.Stats;
+import net.minecraft.tags.ItemTags;
+import net.minecraft.world.entity.ExperienceOrb;
+import net.minecraft.world.entity.item.ItemEntity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.FishingHook;
+import net.minecraft.world.item.FishingRodItem;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.level.storage.loot.BuiltInLootTables;
+import net.minecraft.world.level.storage.loot.LootParams;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.HitResult;
 import net.minecraftforge.api.distmarker.Dist;
@@ -24,11 +38,14 @@ import net.minecraftforge.client.event.RenderLevelStageEvent;
 import net.minecraftforge.event.ItemAttributeModifierEvent;
 import net.minecraftforge.event.entity.living.*;
 import net.minecraftforge.event.entity.player.CriticalHitEvent;
+import net.minecraftforge.event.entity.player.ItemFishedEvent;
 import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.registries.RegistryObject;
+
+import java.util.List;
 
 @Label(name = "New Enchantments", description = "Change some enchantments related stuff and adds new enchantments. Please note that Damaging enchantments such as water coolant are enabled only if ITR 'Replace damaging enchantments' is enabled. This also applies for protection enchantments and ITR 'Replace protection enchantments'")
 @LoadFeature(module = Modules.Ids.EXPERIENCE)
@@ -53,6 +70,7 @@ public class NewEnchantmentsFeature extends Feature {
 	//public static final RegistryObject<Enchantment> CURSE_OF_MENDING = ITERegistries.ENCHANTMENTS.register("curse_of_mending", CurseOfMending::new);
 	public static final RegistryObject<Enchantment> REACH = ITERegistries.ENCHANTMENTS.register("reach", Reach::new);
 	public static final RegistryObject<Enchantment> VINDICATION = ITERegistries.ENCHANTMENTS.register("vindication", Vindication::new);
+	public static final RegistryObject<Enchantment> LUCKY_HOOK = ITERegistries.ENCHANTMENTS.register("lucky_hook", LuckyHook::new);
 	public NewEnchantmentsFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 	}
@@ -94,6 +112,41 @@ public class NewEnchantmentsFeature extends Feature {
 	public void onBreakSpeed(PlayerEvent.BreakSpeed event) {
 		event.setNewSpeed(event.getNewSpeed() + Blasting.getMiningSpeedBoost(event.getEntity(), event.getState()));
 	}
+
+	@SubscribeEvent
+	public void onItemFished(ItemFishedEvent event) {
+		if (event.isCanceled())
+			return;
+		Player player = event.getEntity();
+		boolean hasFishingRod = player.getMainHandItem().getItem() instanceof FishingRodItem;
+		boolean hasFishingRodInOffHand = player.getOffhandItem().getItem() instanceof FishingRodItem;
+		if (!hasFishingRod && !hasFishingRodInOffHand)
+			return;
+		int lvl = hasFishingRod ? player.getMainHandItem().getEnchantmentLevel(LUCKY_HOOK.get()) : player.getOffhandItem().getEnchantmentLevel(LUCKY_HOOK.get());
+		if (lvl == 0 || player.getRandom().nextFloat() > LuckyHook.getChanceToDoubleReel(lvl))
+			return;
+		ItemStack stack = hasFishingRod ? player.getMainHandItem() : player.getOffhandItem();
+		FishingHook hookEntity = event.getHookEntity();
+		LootParams lootparams = (new LootParams.Builder((ServerLevel) player.level())).withParameter(LootContextParams.ORIGIN, hookEntity.position()).withParameter(LootContextParams.TOOL, stack).withParameter(LootContextParams.THIS_ENTITY, hookEntity).withParameter(LootContextParams.KILLER_ENTITY, player).withParameter(LootContextParams.THIS_ENTITY, hookEntity).withLuck((float) hookEntity.luck + player.getLuck()).create(LootContextParamSets.FISHING);
+		LootTable loottable = player.level().getServer().getLootData().getLootTable(BuiltInLootTables.FISHING);
+		List<ItemStack> list = loottable.getRandomItems(lootparams);
+		for(ItemStack itemstack : list) {
+			ItemEntity itementity = new ItemEntity(hookEntity.level(), hookEntity.getX(), hookEntity.getY(), hookEntity.getZ(), itemstack);
+			double xDiff = player.getX() - hookEntity.getX();
+			double yDiff = player.getY() - hookEntity.getY();
+			double zDiff = player.getZ() - hookEntity.getZ();
+			double deltaMovMultiplier = 0.1D;
+			itementity.setDeltaMovement(xDiff * deltaMovMultiplier, yDiff * deltaMovMultiplier + Math.sqrt(Math.sqrt(xDiff * xDiff + yDiff * yDiff + zDiff * zDiff)) * 0.08D, zDiff * deltaMovMultiplier);
+			hookEntity.level().addFreshEntity(itementity);
+			player.level().addFreshEntity(new ExperienceOrb(player.level(), player.getX(), player.getY() + 0.5D, player.getZ() + 0.5D, event.getEntity().getRandom().nextInt(6) + 1));
+			if (itemstack.is(ItemTags.FISHES))
+				player.awardStat(Stats.FISH_CAUGHT, 1);
+		}
+	}
+
+	/*public static boolean doesItemInEitherHandMatches(LivingEntity entity, Predicate<ItemStack> stackPredicate) {
+		return stackPredicate.test(entity.getMainHandItem()) || stackPredicate.test(entity.getOffhandItem());
+	}*/
 
 	@SubscribeEvent
 	public void onExperienceDropped(LivingExperienceDropEvent event) {